feat(train,eval): the BC warm-start entry exists and refuses the wrong artifact, the anchor names its arch, one dtype authority and one parameter identity (AUDIT-1 P8: F-16, F-17, F-19, F-20, F-31, F-32 + F-30)

+++ tests/diagnostics/test_acceptance_witness.py
@@ -96,7 +96,7 @@ def _readout(seed: int) -> dict:
     device = torch.device("cpu")
     engine = LocalInferenceEngine(
         seeded_net(_tiny_arch(), seed=seed).to(device).eval(), device, encoding_spec=spec,
-        fused_graph_caps=None, inference_batching=None, max_in_flight=1,
+        fused_graph_caps=None, inference_batching=None, max_in_flight=1, amp_dtype="bf16",
     )
     try:
         openings = [
