+++ tests/eval/test_rung_seat_off_window.py
@@ -98,7 +98,7 @@ def graph_engine():
     net.eval()
     engine = LocalInferenceEngine(net, torch.device("cpu"), encoding_spec=spec,
                                   fused_graph_caps=_CAPS,
-                                  inference_batching=InferenceBatchingSpec(inference_batch_size=64, inference_max_wait_ms=10), max_in_flight=8)
+                                  inference_batching=InferenceBatchingSpec(inference_batch_size=64, inference_max_wait_ms=10), max_in_flight=8, amp_dtype="bf16")
     try:
         yield engine, spec
     finally:
