+++ tests/selfplay/test_inference_batching_threaded.py
@@ -43,7 +43,7 @@ class _Net(torch.nn.Module):
 def _graph_engine(batching: InferenceBatchingSpec, max_in_flight: int) -> LocalInferenceEngine:
     return LocalInferenceEngine(
         _Net(), _CPU, encoding_spec=_GRAPH_SPEC, fused_graph_caps=_CAPS,
-        inference_batching=batching, max_in_flight=max_in_flight,
+        inference_batching=batching, max_in_flight=max_in_flight, amp_dtype="bf16",
     )
 
 
@@ -105,7 +105,7 @@ def test_a_graph_engine_refuses_an_absent_batching_spec() -> None:
     with pytest.raises(ValueError, match="inference_batching"):
         LocalInferenceEngine(
             _Net(), _CPU, encoding_spec=_GRAPH_SPEC, fused_graph_caps=_CAPS,
-            inference_batching=None, max_in_flight=8,
+            inference_batching=None, max_in_flight=8, amp_dtype="bf16",
         )
 
 
@@ -113,7 +113,7 @@ def test_a_grid_engine_carries_none_and_builds_no_graph_server() -> None:
     """The `None` arm is a real posture, not an oversight: a grid route opens no collector."""
     engine = LocalInferenceEngine(
         _Net(), _CPU, encoding_spec=_GRID_SPEC, fused_graph_caps=None,
-        inference_batching=None, max_in_flight=0,
+        inference_batching=None, max_in_flight=0, amp_dtype="bf16",
     )
     try:
         assert engine._graph_server is None
@@ -160,7 +160,7 @@ def test_the_round_spec_carries_the_batching_across_the_process_seam() -> None:
         result_path="r.json", progress_path="p.txt", ladder_bootstrap_resamples=10,
         ladder_bootstrap_ci_level=0.95, ladder_bootstrap_seed=1,
         ply_cap_adjudication=None, strength_floor=None, fused_graph_caps=_CAPS,
-        inference_batching=batching, leaf_batch_size=8, leaf_build_threads=1,
+        inference_batching=batching, leaf_batch_size=8, amp_dtype="bf16", leaf_build_threads=1,
     )
     back = RoundSpec.from_dict(json.loads(json.dumps(dataclasses.asdict(spec))))
     assert isinstance(back.inference_batching, InferenceBatchingSpec), (
